Also convert subcomponents - DialogContent convert to base + styled component. - DialogFooter convert to base + styled component. - Export new base components
