+++ packages/office-ui-fabric-react/src/components/Dialog/DialogContent.types.ts
@@ -1,18 +1,30 @@
 import * as React from 'react';
-import { DialogContent } from './DialogContent';
+import { DialogContentBase } from './DialogContent.base';
 import { IButtonProps } from '../Button/Button.types';
 import { ResponsiveMode } from '../../utilities/decorators/withResponsiveMode';
+import { IStyle, ITheme } from '../../Styling';
+import { IStyleFunction } from '../../Utilities';
 
 export interface IDialogContent {
 
 }
 
-export interface IDialogContentProps extends React.Props<DialogContent> {
+export interface IDialogContentProps extends React.Props<DialogContentBase> {
   /**
   * Optional callback to access the IDialogContent interface. Use this instead of ref for accessing
   * the public methods and properties of the component.
   */
   componentRef?: (component: IDialogContent) => void;
+
+  /**
+   * Call to provide customized styling that will layer on top of the variant rules
+   */
+  getStyles?: IStyleFunction<IDialogContentStyleProps, IDialogContentStyles>;
+
+  /**
+   * Theme provided by HOC.
+   */
+  theme?: ITheme;
 
   /**
   * Show an 'x' close button in the upper-right corner
@@ -79,4 +91,27 @@ export enum DialogType {
   largeHeader = 1,
   /** Dialog with an 'x' close button in the upper-right corner */
   close = 2
+}
+
+export interface IDialogContentStyleProps {
+  /**
+   * Accept theme prop.
+   */
+  theme: ITheme;
+
+  /**
+   * Accept custom classNames
+   */
+  className?: string;
+
+  // Insert DialogContent style props below
+}
+
+export interface IDialogContentStyles {
+  /**
+   * Style for the root element.
+   */
+  root?: IStyle;
+
+  // Insert DialogContent classNames below
 }
