+++ packages/office-ui-fabric-react/src/components/Dialog/index.ts
@@ -1,6 +1,9 @@
 export * from './Dialog';
 export * from './Dialog.base';
 export * from './DialogContent';
+export * from './DialogContent.base';
 export * from './DialogFooter';
+export * from './DialogFooter.base';
 export * from './Dialog.types';
 export * from './DialogContent.types';
+export * from './DialogFooter.types';
